+++ routes/web.php
@@ -6,8 +6,8 @@ $router->get('/contact', 'app/controllers/contact.php');
 $router->get('/notes', 'app/controllers/notes/index.php');
 
 $router->get('/note', 'app/controllers/notes/show.php');
-$router->delete('/note', 'app/controllers/notes/show.php');
+$router->delete('/note', 'app/controllers/notes/destroy.php');
 
 $router->get('/note-create', 'app/controllers/notes/create.php');
-$router->post('/note-create', 'app/controllers/notes/create.php');
+$router->post('/note-create', 'app/controllers/notes/store.php');
